AUDIO WIP refactoring registers
